pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["sasuke13 <bgnydyn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
flake8 = "^6.1.0"
celery = {extras = ["redis"], version = "^5.3.5"}
django-celery-beat = "^2.5.0"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.3.0"
uuid = "^1.30"
auto-dataclass = "^0.1.6"
django-annoying = "^0.10.6"
dependency-injector = "^4.41.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
